+++ helen/__init__.py
@@ -1,3 +1,3 @@
 """helen package."""
 
-__version__ = "1.40.0"
+__version__ = "1.41.0"